advanced-version/services/api/requirements.txt
flask==2.0.0
flask-httpauth==4.1.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.5
python-dotenv==0.14.0
Werkzeug==2.0.0
